Apply initial body diff earlier, on DOMContentLoaded
`DOMContentLoaded` is the earliest event at which point Reactize can be
certain the document is ready to reactify. `load` waits for all assets
to download, but that is not necessary.

This leads to expected behavior if developers attach event listeners in
`DOMContentLoaded` (as long as their DOMCL listener is attached after
Reactive is included).

Addresses #17.